+++ config/dev.exs
@@ -3,6 +3,8 @@ use Mix.Config
 # Do not include metadata nor timestamps in development logs
 config :logger, :console, format: "[$level] $message\n"
 
+config :mix_test_watch, clear: true
+
 config :eventstore, EventStore.Storage,
   serializer: EventStore.TermSerializer,
   username: "postgres",