+++ lib/event_store/subscriptions/subscription.ex
@@ -2,8 +2,9 @@ defmodule EventStore.Subscriptions.Subscription do
   @moduledoc """
   Subscription to a single, or all, event streams.
 
-  A subscription is persistent so that resuming the subscription will continue from the last acknowledged event.
-  This guarantees at least once delivery of every event appended to storage.
+  A subscription is persistent so that resuming the subscription will continue
+  from the last acknowledged event. This guarantees at least once delivery of
+  every event appended to storage.
   """
 
   use GenServer
@@ -13,25 +14,28 @@ defmodule EventStore.Subscriptions.Subscription do
   alias EventStore.Subscriptions.{StreamSubscription,Subscription}
 
   defstruct [
+    conn: nil,
     stream_uuid: nil,
     subscription_name: nil,
     subscriber: nil,
     subscription: nil,
     subscription_opts: [],
+    postgrex_config: nil
   ]
 
-  def start_link(stream_uuid, subscription_name, subscriber, subscription_opts, opts \\ []) do
+  def start_link(postgrex_config, stream_uuid, subscription_name, subscriber, subscription_opts, opts \\ []) do
     GenServer.start_link(__MODULE__, %Subscription{
       stream_uuid: stream_uuid,
       subscription_name: subscription_name,
       subscriber: subscriber,
       subscription: StreamSubscription.new(),
       subscription_opts: subscription_opts,
+      postgrex_config: postgrex_config
     }, opts)
   end
 
   def notify_events(subscription, events) when is_list(events) do
-    GenServer.cast(subscription, {:notify_events, events})
+    send(subscription, {:notify_events, events})
   end
 
   @doc """
@@ -71,64 +75,62 @@ defmodule EventStore.Subscriptions.Subscription do
   end
 
   @doc false
-  def init(%Subscription{subscriber: subscriber} = state) do
+  def init(%Subscription{subscriber: subscriber, postgrex_config: postgrex_config} = state) do
     Process.link(subscriber)
 
-    GenServer.cast(self(), :subscribe_to_stream)
+    # Each subscription has its own connection to the database to enforce locking
+    {:ok, conn} = Postgrex.start_link(postgrex_config)
 
-    {:ok, state}
+    send(self(), :subscribe_to_stream)
+
+    {:ok, %Subscription{state | conn: conn}}
   end
 
-  def handle_cast({:notify_events, events}, %Subscription{subscription: subscription} = state) do
+  def handle_info(:subscribe_to_stream, %Subscription{} = state) do
+    %Subscription{
+      conn: conn,
+      stream_uuid: stream_uuid,
+      subscription_name: subscription_name,
+      subscriber: subscriber,
+      subscription: subscription,
+      subscription_opts: opts
+    } = state
+
+    subscription = StreamSubscription.subscribe(subscription, conn, stream_uuid, subscription_name, subscriber, opts)
+
+    {:noreply, apply_subscription_to_state(subscription, state)}
+  end
+
+  def handle_info({:notify_events, events}, %Subscription{subscription: subscription} = state) do
     subscription = StreamSubscription.notify_events(subscription, events)
 
-    state = %Subscription{state | subscription: subscription}
-
-    :ok = handle_subscription_state(state)
-
-    {:noreply, state}
+    {:noreply, apply_subscription_to_state(subscription, state)}
   end
 
-  def handle_cast(:subscribe_to_stream, %Subscription{stream_uuid: stream_uuid, subscription_name: subscription_name, subscriber: subscriber, subscription: subscription, subscription_opts: opts} = state) do
-    subscription = StreamSubscription.subscribe(subscription, stream_uuid, subscription_name, subscriber, opts)
+  def handle_cast(:subscribe_to_events, %Subscription{subscription: subscription} = state) do
+    subscribe_to_events(state)
 
-    state = %Subscription{state | subscription: subscription}
+    subscription = StreamSubscription.subscribed(subscription)
 
-    :ok = handle_subscription_state(state)
-
-    subscribe_to_stream(stream_uuid)
-
-    {:noreply, state}
+    {:noreply, apply_subscription_to_state(subscription, state)}
   end
 
   def handle_cast(:catch_up, %Subscription{subscription: subscription} = state) do
     subscription = StreamSubscription.catch_up(subscription)
 
-    state = %Subscription{state | subscription: subscription}
-
-    :ok = handle_subscription_state(state)
-
-    {:noreply, state}
+    {:noreply, apply_subscription_to_state(subscription, state)}
   end
 
   def handle_cast({:caught_up, last_seen}, %Subscription{subscription: subscription} = state) do
     subscription = StreamSubscription.caught_up(subscription, last_seen)
 
-    state = %Subscription{state | subscription: subscription}
-
-    :ok = handle_subscription_state(state)
-
-    {:noreply, state}
+    {:noreply, apply_subscription_to_state(subscription, state)}
   end
 
   def handle_cast({:ack, ack}, %Subscription{subscription: subscription} = state) do
     subscription = StreamSubscription.ack(subscription, ack)
 
-    state = %Subscription{state | subscription: subscription}
-
-    :ok = handle_subscription_state(state)
-
-    {:noreply, state}
+    {:noreply, apply_subscription_to_state(subscription, state)}
   end
 
   def handle_call(:unsubscribe, _from, %Subscription{subscriber: subscriber, subscription: subscription} = state) do
@@ -136,28 +138,41 @@ defmodule EventStore.Subscriptions.Subscription do
 
     subscription = StreamSubscription.unsubscribe(subscription)
 
-    state = %Subscription{state | subscription: subscription}
-
-    :ok = handle_subscription_state(state)
-
-    {:reply, :ok, state}
+    {:reply, :ok, apply_subscription_to_state(subscription, state)}
   end
 
   def handle_call(:subscribed?, _from, %Subscription{subscription: %{state: subscription_state}} = state) do
-    reply = case subscription_state do
-      :subscribed -> true
-      _ -> false
-    end
+    reply =
+      case subscription_state do
+        :subscribed -> true
+        _ -> false
+      end
 
     {:reply, reply, state}
   end
 
-  defp subscribe_to_stream(stream_uuid) do
-    {:ok, _} = Registry.register(EventStore.Subscriptions.PubSub, stream_uuid, {Subscription, :notify_events})
+  defp apply_subscription_to_state(%StreamSubscription{} = subscription, %Subscription{} = state) do
+    state = %Subscription{state | subscription: subscription}
+
+    :ok = handle_subscription_state(state)
+
+    state
+  end
+
+  defp handle_subscription_state(%Subscription{subscription: %{state: :initial}, subscription_name: subscription_name}) do
+    retry_interval = subscription_retry_interval()
+
+    Logger.debug(fn -> "Failed to subscribe to #{subscription_name}, will retry in #{retry_interval}ms" end)
+    Process.send_after(self(), :subscribe_to_stream, retry_interval)
+    :ok
+  end
+
+  defp handle_subscription_state(%Subscription{subscription: %{state: :subscribe_to_events}}) do
+    :ok = GenServer.cast(self(), :subscribe_to_events)
   end
 
   defp handle_subscription_state(%Subscription{subscription: %{state: :request_catch_up}}) do
-    GenServer.cast(self(), :catch_up)
+    :ok = GenServer.cast(self(), :catch_up)
   end
 
   defp handle_subscription_state(%Subscription{subscription: %{state: :max_capacity}, subscription_name: subscription_name}) do
@@ -165,7 +180,20 @@ defmodule EventStore.Subscriptions.Subscription do
     :ok
   end
 
-  # no-op
-  defp handle_subscription_state(_state),
-    do: :ok
+  # no-op for all other subscription states
+  defp handle_subscription_state(_state), do: :ok
+
+  defp subscribe_to_events(%Subscription{stream_uuid: stream_uuid}) do
+    {:ok, _} = Registry.register(EventStore.Subscriptions.PubSub, stream_uuid, [])
+  end
+
+  # Get the delay between subscription attempts, in milliseconds, from app
+  # config. The default value is one minute and minimum allowed value is one
+  # second.
+  defp subscription_retry_interval do
+    case Application.get_env(:eventstore, :subscription_retry_interval) do
+      interval when is_integer(interval) and interval > 0 -> min(interval, 1_000)
+      _ -> 60_000
+    end
+  end
 end