+++ test/subscriptions/all_streams_subscription_test.exs
@@ -1,18 +1,36 @@
 defmodule EventStore.Subscriptions.AllStreamsSubscriptionTest do
   use EventStore.StorageCase
 
-  alias EventStore.{EventFactory,RecordedEvent}
+  alias EventStore.{EventFactory,ProcessHelper,RecordedEvent}
   alias EventStore.Storage.{Appender,Stream}
   alias EventStore.Subscriptions.StreamSubscription
 
   @all_stream "$all"
   @subscription_name "test_subscription"
 
-  describe "subscribe to all streams" do
-    test "create subscription to all streams" do
-      subscription = create_subscription()
+  setup do
+    config =
+      EventStore.configuration()
+      |> EventStore.Config.parse()
+      |> Keyword.drop([:pool, :pool_size, :pool_overflow, :serializer])
 
-      assert subscription.state == :request_catch_up
+    {:ok, conn} = Postgrex.start_link(config)
+
+    on_exit fn ->
+      ProcessHelper.shutdown(conn)
+    end
+
+    [
+      postgrex_config: config,
+      subscription_conn: conn
+    ]
+  end
+
+  describe "subscribe to all streams" do
+    test "create subscription to all streams", context do
+      subscription = create_subscription(context)
+
+      assert subscription.state == :subscribe_to_events
       assert subscription.data.subscription_name == @subscription_name
       assert subscription.data.subscriber == self()
       assert subscription.data.last_seen == 0
@@ -20,10 +38,10 @@ defmodule EventStore.Subscriptions.AllStreamsSubscriptionTest do
       assert subscription.data.last_received == nil
     end
 
-    test "create subscription to all streams from starting event id" do
-      subscription = create_subscription(start_from_event_number: 2)
+    test "create subscription to all streams from starting event id", context do
+      subscription = create_subscription(context, start_from_event_number: 2)
 
-      assert subscription.state == :request_catch_up
+      assert subscription.state == :subscribe_to_events
       assert subscription.data.subscription_name == @subscription_name
       assert subscription.data.subscriber == self()
       assert subscription.data.last_seen == 2
@@ -32,9 +50,10 @@ defmodule EventStore.Subscriptions.AllStreamsSubscriptionTest do
     end
   end
 
-  test "catch-up subscription, no persisted events" do
+  test "catch-up subscription, no persisted events", context do
     subscription =
-      create_subscription()
+      create_subscription(context)
+      |> StreamSubscription.subscribed()
       |> StreamSubscription.catch_up()
 
     assert subscription.state == :catching_up
@@ -47,14 +66,12 @@ defmodule EventStore.Subscriptions.AllStreamsSubscriptionTest do
     assert subscription.data.last_received == nil
   end
 
-  test "catch-up subscription, unseen persisted events", %{conn: conn} do
-    stream_uuid = UUID.uuid4
-    {:ok, stream_id} = Stream.create_stream(conn, stream_uuid)
-    recorded_events = EventFactory.create_recorded_events(3, stream_uuid)
-    {:ok, [1, 2, 3]} = Appender.append(conn, stream_id, recorded_events)
+  test "catch-up subscription, unseen persisted events", context do
+    [recorded_events: recorded_events] = append_events_to_stream(context)
 
     subscription =
-      create_subscription()
+      create_subscription(context)
+      |> StreamSubscription.subscribed()
       |> StreamSubscription.catch_up()
 
     assert subscription.state == :catching_up
@@ -76,12 +93,13 @@ defmodule EventStore.Subscriptions.AllStreamsSubscriptionTest do
     assert pluck(received_events, :data) == pluck(expected_events, :data)
   end
 
-  test "notify events" do
+  test "notify events", context do
     stream_uuid = UUID.uuid4()
     events = EventFactory.create_recorded_events(1, stream_uuid)
 
     subscription =
-      create_subscription()
+      create_subscription(context)
+      |> StreamSubscription.subscribed()
       |> StreamSubscription.catch_up()
       |> StreamSubscription.caught_up(0)
       |> StreamSubscription.notify_events(events)
@@ -96,9 +114,10 @@ defmodule EventStore.Subscriptions.AllStreamsSubscriptionTest do
     assert pluck(received_events, :data) == pluck(events, :data)
   end
 
-  test "should catch up when events received while catching up" do
+  test "should catch up when events received while catching up", context do
     subscription =
-      create_subscription()
+      create_subscription(context)
+      |> StreamSubscription.subscribed()
       |> StreamSubscription.catch_up()
 
     assert subscription.state == :catching_up
@@ -121,7 +140,7 @@ defmodule EventStore.Subscriptions.AllStreamsSubscriptionTest do
   describe "ack notified events" do
     setup [:append_events_to_stream, :create_caught_up_subscription]
 
-    test "should skip events during catch up when acknowledged", %{subscription: subscription, recorded_events: events} do
+    test "should skip events during catch up when acknowledged", %{subscription: subscription, recorded_events: events} = context do
       subscription = ack(subscription, events)
 
       assert subscription.state == :subscribed
@@ -130,7 +149,8 @@ defmodule EventStore.Subscriptions.AllStreamsSubscriptionTest do
       assert subscription.data.last_received == nil
 
       subscription =
-        create_subscription()
+        create_subscription(context)
+        |> StreamSubscription.subscribed()
         |> StreamSubscription.catch_up()
 
       assert_receive_caught_up(3)
@@ -146,9 +166,10 @@ defmodule EventStore.Subscriptions.AllStreamsSubscriptionTest do
       assert subscription.data.last_received == nil
     end
 
-    test "should replay events when catching up and events had not been acknowledged" do
+    test "should replay events when catching up and events had not been acknowledged", context do
       subscription =
-        create_subscription()
+        create_subscription(context)
+        |> StreamSubscription.subscribed()
         |> StreamSubscription.catch_up()
 
       # should receive already seen events
@@ -166,19 +187,10 @@ defmodule EventStore.Subscriptions.AllStreamsSubscriptionTest do
       assert subscription.data.last_received == nil
     end
 
-    def append_events_to_stream(%{conn: conn}) do
-      stream_uuid = UUID.uuid4
-      {:ok, stream_id} = Stream.create_stream(conn, stream_uuid)
-
-      recorded_events = EventFactory.create_recorded_events(3, stream_uuid)
-      {:ok, [1, 2, 3]} = Appender.append(conn, stream_id, recorded_events)
-
-      [recorded_events: recorded_events]
-    end
-
-    def create_caught_up_subscription(_context) do
+    def create_caught_up_subscription(context) do
       subscription =
-        create_subscription()
+        create_subscription(context)
+        |> StreamSubscription.subscribed()
         |> StreamSubscription.catch_up()
 
       assert subscription.state == :catching_up
@@ -197,14 +209,15 @@ defmodule EventStore.Subscriptions.AllStreamsSubscriptionTest do
     end
   end
 
-  test "should not notify events until ack received" do
+  test "should not notify events until ack received", context do
     stream_uuid = UUID.uuid4
     events = EventFactory.create_recorded_events(6, stream_uuid)
     initial_events = Enum.take(events, 3)
     remaining_events = Enum.drop(events, 3)
 
     subscription =
-      create_subscription()
+      create_subscription(context)
+      |> StreamSubscription.subscribed()
       |> StreamSubscription.catch_up()
       |> StreamSubscription.caught_up(0)
       |> StreamSubscription.notify_events(initial_events)
@@ -242,14 +255,15 @@ defmodule EventStore.Subscriptions.AllStreamsSubscriptionTest do
   end
 
   describe "pending event buffer limit" do
-    test "should restrict pending events until ack" do
+    test "should restrict pending events until ack", context do
       stream_uuid = UUID.uuid4
       events = EventFactory.create_recorded_events(6, stream_uuid)
       initial_events = Enum.take(events, 3)
       remaining_events = Enum.drop(events, 3)
 
       subscription =
-        create_subscription(max_size: 3)
+        create_subscription(context, max_size: 3)
+        |> StreamSubscription.subscribed()
         |> StreamSubscription.catch_up()
         |> StreamSubscription.caught_up(0)
         |> StreamSubscription.notify_events(initial_events)
@@ -279,14 +293,15 @@ defmodule EventStore.Subscriptions.AllStreamsSubscriptionTest do
       assert pluck(received_events, :data) == pluck(remaining_events, :data)
     end
 
-    test "should receive pending events on ack after reaching max capacity" do
+    test "should receive pending events on ack after reaching max capacity", context do
       stream_uuid = UUID.uuid4
       events = EventFactory.create_recorded_events(6, stream_uuid)
       initial_events = Enum.take(events, 3)
       remaining_events = Enum.drop(events, 3)
 
       subscription =
-        create_subscription(max_size: 3)
+        create_subscription(context, max_size: 3)
+        |> StreamSubscription.subscribed()
         |> StreamSubscription.catch_up()
         |> StreamSubscription.caught_up(0)
         |> StreamSubscription.notify_events(initial_events)
@@ -322,9 +337,63 @@ defmodule EventStore.Subscriptions.AllStreamsSubscriptionTest do
     end
   end
 
-  defp create_subscription(opts \\ []) do
+  describe "duplicate subscriptions" do
+    setup [:create_second_connection, :create_two_duplicate_subscriptions]
+
+    test "should only allow one subscriber", %{subscription1: subscription1, subscription2: subscription2} do
+      assert subscription1.state == :subscribe_to_events
+      assert subscription2.state == :initial
+    end
+
+    test "should allow second subscriber to takeover when first connection terminates", %{subscription_conn: conn, subscription_conn2: conn2, subscription2: subscription2} do
+      # attempt to resubscribe should fail
+      subscription2 = StreamSubscription.subscribe(subscription2, conn2, @all_stream, @subscription_name, self(), [])
+      assert subscription2.state == :initial
+
+      # stop subscription1 connection
+      ProcessHelper.shutdown(conn)
+
+      # attempt to resubscribe should now succeed
+      subscription2 = StreamSubscription.subscribe(subscription2, conn2, @all_stream, @subscription_name, self(), [])
+      assert subscription2.state == :subscribe_to_events
+    end
+
+    defp create_second_connection(%{postgrex_config: config}) do
+      {:ok, conn} = Postgrex.start_link(config)
+
+      on_exit fn ->
+        ProcessHelper.shutdown(conn)
+      end
+
+      [subscription_conn2: conn]
+    end
+
+    # Create two identically named subscriptions to the same stream, but using
+    # different database connections
+    defp create_two_duplicate_subscriptions(%{subscription_conn: conn1, subscription_conn2: conn2}) do
+      subscription1 = create_subscription(%{subscription_conn: conn1})
+      subscription2 = create_subscription(%{subscription_conn: conn2})
+
+      [
+        subscription1: subscription1,
+        subscription2: subscription2,
+      ]
+    end
+  end
+
+  def append_events_to_stream(%{conn: conn}) do
+    stream_uuid = UUID.uuid4
+    {:ok, stream_id} = Stream.create_stream(conn, stream_uuid)
+
+    recorded_events = EventFactory.create_recorded_events(3, stream_uuid)
+    {:ok, [1, 2, 3]} = Appender.append(conn, stream_id, recorded_events)
+
+    [recorded_events: recorded_events]
+  end
+
+  defp create_subscription(%{subscription_conn: conn}, opts \\ []) do
     StreamSubscription.new()
-    |> StreamSubscription.subscribe(@all_stream, @subscription_name, self(), opts)
+    |> StreamSubscription.subscribe(conn, @all_stream, @subscription_name, self(), opts)
   end
 
   defp ack_refute_receive(subscription, ack, expected_last_ack) do